Fixing multiplication result converted to larger type reports in incremental tests (#7454)
* Fixing multiplication result converted to larger type reports in incremental tests

Fix https://github.com/kokkos/kokkos/security/code-scanning/17
https://github.com/kokkos/kokkos/security/code-scanning/16
https://github.com/kokkos/kokkos/security/code-scanning/15
https://github.com/kokkos/kokkos/security/code-scanning/14
https://github.com/kokkos/kokkos/security/code-scanning/13
https://github.com/kokkos/kokkos/security/code-scanning/12
https://github.com/kokkos/kokkos/security/code-scanning/11
https://github.com/kokkos/kokkos/security/code-scanning/10

* Fixup avoid integer overflow

+++ core/unit_test/incremental/Test12a_ThreadScratch.hpp
@@ -88,7 +88,7 @@ struct ThreadScratch {
     auto v_H = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), v);
 
     size_t check   = 0;
-    const size_t s = pN * sX * sY;
+    const size_t s = static_cast<size_t>(pN) * sX * sY;
     for (int n = 0; n < pN; ++n)
       for (int m = 0; m < max_team_size; ++m) {
         check += v_H(n, m);
